requirements.txt
discord.py
twilio
python-dotenv
urllib.request
bs4
requests
python-espncricinfo
lxml